Initial commit of Duel game.
Name is probably taken, but it will be fine for a working title.
Game play is basic: Two ships in orbit around a sun. You try to
lob torpedos at each other around the sun. You have thrusters to
get into a better position, but you might just run yourself into
the sun. so be careful with thruster use!

I'm using the kivy Joystick I made the other day. I've already
found a problem with it. Two joysticks at the same time doesn't
work very well. I need them to track which touch they started from.
Mostly the issue is that on_touch_move doesn't check the touch it
recieved to see if it is the same touch as previous.