Trails and Tuuuuuuuuuuuuuuubes!
I have added trails to the ships and torpedoes. It makes the game
much prettier and gives a better sense of movement seeing the trail
of the objects flying around. It makes the game feel like it is
actually a legit game and not some high school project.

Torpedoes now collide with each other, so you can theoretically
protect yourself from an incoming torpedo, though it would be
pretty spectacular if you did. I imagine it could work well for
spray and pray on an incoming torpedo.

The system for firing torpedoes is greatly simplified. But I did it
in a way that should allow for some fun later. I've got it set now
that you have one fire button and one recall button. This allows
you to just have a panic button that kills your own torpedoes
should you be about to kill yourself. The number of torpedoes will
also be something that can be edited.

I've made a burn down chart. 11 items on it before I can make an
initial release of the game onto the android play store. I hope to
do so in just a few more weeks like this one.