Slow progress tonight.
I did not get as much done as I had hoped I would. I have a managed
to get the game seperated out into screens now. That helps. You can
start a new game and it will throw up a victory screen and then
you can go back to the menu or play again. Progress to be sure.

What I'm slightly annoyed with is my lack of progress on the
settings front. I was trying to get the settings screen to work
and I just didn't quite get it because I couldn't figure out how
to get config to behave. And I need to go empty the litter box now
before going to bed. Hopefully I'll be able to accomplish more
later this week.