I forgot to commit last time.
I got so carried away with the debugging last time that I forgot
to commit. woops. oh well. So what have I done tonight?

Well, tonight specifically I haven't done too much. I got the
settings to work properly. This is good as it allows the game to be
playable again on handhelds. It wasn't work for a while as a result
of the config not finding necessary values.

Last time I coded I accomplished quite a number of things. I got
the settings screen up and running, if not fully working. I fixed
the issue where red wouldn't be able to fire until they recalled
their torpedoes at the begining of the game. Shots that collide or
fall into the sun now get removed from the counter. When you start
a new game, the previous shots are removed, clearing the gamefield.
And I added a score keeper. I'm intending to add a "first to x"
game mode in the rules, so this puts down one of the needed
components of that.

Left to do:
Add an icon.
First to x matches
ship selector
nicer ships
explosion on ship destroyed
sound

which honestly isn't a particularly long list! Hopefully I'll have
something up in the app store here soon. w00t w00t!