oooh, two commits in one night!
I added a makefile. For some reason buildozer won't compile the apk
on my NFS drive. I must copy the code to my physical drive and then
it will build properly. So the make file is just a save of that set
of copy functions. And then some where to remember the actual build
call to buildozer.