EXPLOSIONS!
I have accomplished two things tonight. First of is the icon. I now
have a place holder icon replacing the kivy icon for the apk. I'll
come up with something better later or have Kennan come up with
something. We'll see. But for now it is better than the default.
The other thing was explosions! Which actually has two facets.
First: A time delay when your ship gets destroyed. This allows you
a moment to realize that you've died. But further, this gives a
moment for debris to fly away! It's pretty forceful, espectially if
you crash into the sun.

Later I will make use of this time delay for allowing ties. It will
take a little effort since it will require a rework of the ending
code. But it shouldn't be to annoyingly challenging.