The one that is trown away.
I'm drawing this project to a close for now. With this version I
have actually released the game to the play store. I really want to
recreate this with networking in mind. Make it a multi-player
multi-device game. But that will basically require me to rewrite
the entirity of the base logic to get it to work. To quote MM-M:

    Plan to throw one away.

This would be that. It's been good, but it's time has come. I've
learned a lot in this initial push, but it is time to move on to
a different project to continue my learnings.